Implemented the new date system where each ticker can have it's own number of bars
The system no longer depends on every ticker having the same number of
bars. Within the best of strategy they all have the same number of bars
because its all on the same ticker. But the main strategy doesn't care
anymore. It just gets whatever tickers are trading on that day.